pyproject.toml
[project]
name = "vujudge"
version = "0.1.0"
description = "VUJudge is an online judge platform developed for the Varendra University Programming Club (VUPC)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-allauth-ui>=1.5.0",
    "django-allauth>=65.0.2",
    "django-unfold>=0.40.0",
    "django-widget-tweaks>=1.5.0",
    "django>=5.1.2",
    "httpx>=0.27.2",
    "slippers>=0.6.2",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.9",
]
